Adds integration tests (#33)
The testing strategy here is simply to include tests that exercise each
order grouping at least once. It is expected that implementation
features should be tested in the respective implementation buildpack.

Since there is only one group so far in this buildpack, there is only a
single test to ensure it detects and builds correctly.