Switch puma to webrick
It looks like Ruby stopped bundling webrick, so when we went from 3.1 to 3.3 it failed. I tried adding in Puma, but then the Puma buildpack runs and that breaks the rackup tests, because the rackup buildpack doesn't run. I am putting webrick now, as that is what was previously bundled in Ruby 3.1.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>